Move Brand to root
