+++ homewizard_energy/v2/const.py
@@ -1,6 +1,1 @@
-"""Constants for HomeWizard Energy."""
-
-SUPPORTED_API_VERSION = "v1"
-
-SUPPORTS_STATE = ["HWE-SKT"]
-SUPPORTS_IDENTIFY = ["HWE-SKT", "HWE-P1", "HWE-WTR"]
+"""Constants for HomeWizard Energy v2."""
